Move FieldLabelKind
