+++ lib/cdc/util/optional_schema.go
@@ -3,12 +3,13 @@ package util
 import (
 	"log/slog"
 
-	"github.com/artie-labs/transfer/lib/cdc"
+	"github.com/artie-labs/transfer/lib/debezium"
+
 	"github.com/artie-labs/transfer/lib/typing"
 )
 
 func (s *SchemaEventPayload) GetOptionalSchema() map[string]typing.KindDetails {
-	fieldsObject := s.Schema.GetSchemaFromLabel(cdc.After)
+	fieldsObject := s.Schema.GetSchemaFromLabel(debezium.After)
 	if fieldsObject == nil {
 		// AFTER schema does not exist.
 		return nil
