+++ lib/debezium/types.go
@@ -15,6 +15,18 @@ import (
 	"github.com/artie-labs/transfer/lib/typing/ext"
 )
 
+// FieldLabelKind is used when the schema is turned on. Each schema object will be labelled.
+type FieldLabelKind string
+
+const (
+	Before      FieldLabelKind = "before"
+	After       FieldLabelKind = "after"
+	Source      FieldLabelKind = "source"
+	Op          FieldLabelKind = "op"
+	TsMs        FieldLabelKind = "ts_ms"
+	Transaction FieldLabelKind = "transaction"
+)
+
 type SupportedDebeziumType string
 
 const (
